fix: Address bot review feedback on PR #17
- P1: LoadProject/ListProjects now catch JsonException, IOException,
  UnauthorizedAccessException and log warnings instead of crashing
- P2: Clear stale raster state when loading project with empty path
- Remove unused IOptions<ProjectStorageOptions> inject from razor
- Thread-safe: assign _rasterInfo after await, not inside Task.Run
- BandStatistics: use IReadOnlyList<long> for histogram (no overflow)
- RasterService: histogram uses long[] bins for large rasters
- Reserved name validation catches names with extensions (CON.txt)

+++ src/FuzzySat.Web/Services/RasterService.cs
@@ -61,8 +61,8 @@ public sealed class RasterService
         }
         var stdDev = Math.Sqrt(sumSqDiff / count);
 
-        // Histogram (256 bins)
-        var histogram = new int[256];
+        // Histogram (256 bins, long to avoid overflow on large rasters)
+        var histogram = new long[256];
         var range = max - min;
         if (range > 0)
         {
@@ -79,7 +79,7 @@ public sealed class RasterService
         }
         else
         {
-            histogram[128] = (int)count;
+            histogram[128] = count;
         }
 
         return new BandStatistics(min, max, mean, stdDev, histogram);
